convert TitleText.js to TS
